+++ static/js/index.js
@@ -1,5 +1,6 @@
 var L = require('leaflet'),
     d3 = require('d3'),
+    summary = require('./summary').summary,
     mapa = L.map('mapaleaf', {
         attributionControl: false,
         zoomControl: false
@@ -53,39 +54,12 @@ function plot(url){
 }
 
 d3.json('/incidents/summary', function(json){
-    var data = d3.entries(json).sort(function(a, b){
-        return b.value - a.value;
-    }),
-    max = d3.max(data, function(d){ return d.value;}),
-    summary = d3.select('div.boxes')
-                .append('div')
-                .attr('id', 'summary')
-                .append('div')
-                .selectAll('div')
-                .data(data)
-                .enter()
-                .append('div');
-
-    function slug(d){
-        return d.key.split(' ').join('');
-    }
-
-    summary.append('input')
-        .attr('id', slug)
-        .attr('type', 'radio')
-        .attr('name', 'finalKind')
+    d3.select('div.boxes')
+        .datum(json)
+        .call(summary())
+        .selectAll('input')
         .on('change', function(value){
             plot('/incidents?finalKind=' + value.key);
         });
-
-    summary.append('label')
-        .attr('class', slug)
-        .attr('for', slug)
-        .append('span')
-        .attr('class', 'bar')
-        .style('width', function(d){
-            return ((d.value / max) * 100) + '%';
-        })
-        .text(function(d){ return d.key;});
 });
